+++ zomes/tests/utils.ts
@@ -1,0 +1,14 @@
+import faker from "faker"
+
+export function generate_link_expression(agent: string) {
+    return {
+      data: {source: faker.name.findName(), target: faker.name.findName(), predicate: faker.name.findName()},
+      author: agent, 
+      timestamp: new Date().toISOString(), 
+      proof: {signature: "sig", key: "key"},
+   }
+}
+
+export function sleep(ms) {
+    return new Promise(resolve => setTimeout(resolve, ms));
+}
